Integrate PopularMovies page with API

+++ app/src/main/java/com/example/movieapplication/model/Movies.kt
@@ -22,6 +22,6 @@ data class Movie (
     @SerializedName("release_date") val releaseDate : String,
     @SerializedName("title") val title : String,
     @SerializedName("video") val video : Boolean,
-    @SerializedName("vote_average") val voteAverage : Int,
+    @SerializedName("vote_average") val voteAverage : Double,
     @SerializedName("vote_count") val voteCount : Int
 )
